# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributor Pricing & Stock Comparison API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  createdAt: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Distributor {
  id: number;
  name: string;
  isBaseline: boolean;
  stalenessThresholdDays: number;
  createdAt: string;
  /** @nullable */
  lastUploadAt?: string | null;
  /** @nullable */
  lastUploadStatus?: string | null;
}

export interface DistributorInput {
  /** @minLength 1 */
  name: string;
  isBaseline?: boolean;
  stalenessThresholdDays?: number;
}

export interface DistributorUpdate {
  /** @minLength 1 */
  name?: string;
  isBaseline?: boolean;
  stalenessThresholdDays?: number;
}

export interface Brand {
  id: number;
  canonicalName: string;
  aliases: string[];
}

export interface BrandInput {
  /** @minLength 1 */
  canonicalName: string;
  aliases: string[];
}

export interface BrandUpdate {
  /** @minLength 1 */
  canonicalName?: string;
  aliases?: string[];
}

export interface ColumnMapping {
  vpn: string;
  brand: string;
  description: string;
  sell_price: string;
  soh: string;
  /** @nullable */
  soo?: string | null;
}

export type ImportProfileSourceFormat = typeof ImportProfileSourceFormat[keyof typeof ImportProfileSourceFormat];


export const ImportProfileSourceFormat = {
  xlsx: 'xlsx',
  txt: 'txt',
} as const;

export interface ImportProfile {
  id: number;
  distributorId: number;
  sourceFormat: ImportProfileSourceFormat;
  /** @nullable */
  delimiter?: string | null;
  headerRowIndex: number;
  mapping: ColumnMapping;
  createdAt: string;
  updatedAt: string;
}

export type ImportProfileInputSourceFormat = typeof ImportProfileInputSourceFormat[keyof typeof ImportProfileInputSourceFormat];


export const ImportProfileInputSourceFormat = {
  xlsx: 'xlsx',
  txt: 'txt',
} as const;

export interface ImportProfileInput {
  sourceFormat: ImportProfileInputSourceFormat;
  /** @nullable */
  delimiter?: string | null;
  headerRowIndex: number;
  mapping: ColumnMapping;
}

export type UploadStatus = typeof UploadStatus[keyof typeof UploadStatus];


export const UploadStatus = {
  parsing: 'parsing',
  mapped: 'mapped',
  committed: 'committed',
  failed: 'failed',
} as const;

export interface Upload {
  id: number;
  distributorId: number;
  distributorName?: string;
  filename: string;
  uploadedAt: string;
  snapshotDate: string;
  rowCountTotal: number;
  rowCountMatched: number;
  status: UploadStatus;
}

export interface ParseUploadInput {
  distributorId: number;
  headerRowIndex?: number;
  delimiter?: string;
}

export type ParsePreviewRowsItem = { [key: string]: unknown };

export interface ParsePreview {
  tempFileKey: string;
  columns: string[];
  rows: ParsePreviewRowsItem[];
  rowCountTotal: number;
  rowCountMatched: number;
  hasProfile: boolean;
  /** @nullable */
  detectedDelimiter?: string | null;
  profile?: ImportProfile;
}

export type CommitUploadInputSourceFormat = typeof CommitUploadInputSourceFormat[keyof typeof CommitUploadInputSourceFormat];


export const CommitUploadInputSourceFormat = {
  xlsx: 'xlsx',
  txt: 'txt',
} as const;

export interface CommitUploadInput {
  distributorId: number;
  tempFileKey: string;
  mapping: ColumnMapping;
  sourceFormat: CommitUploadInputSourceFormat;
  /** @nullable */
  delimiter?: string | null;
  headerRowIndex: number;
  snapshotDate?: string;
  saveProfile?: boolean;
}

export interface DistributorStockEntry {
  distributorId: number;
  distributorName: string;
  isBaseline: boolean;
  /** @nullable */
  sellPrice?: number | null;
  /** @nullable */
  soh?: number | null;
  /** @nullable */
  soo?: number | null;
  /** @nullable */
  movement?: number | null;
  /** @nullable */
  movementSinceDate?: string | null;
  isNew?: boolean;
  /** @nullable */
  priceDelta?: number | null;
  /** @nullable */
  priceDeltaPct?: number | null;
}

export interface ComparisonRow {
  productId: number;
  vpnNormalized: string;
  vpnDisplay: string;
  brand: string;
  description: string;
  distributors: DistributorStockEntry[];
  /** @nullable */
  cheapestCompetitorId: number | null;
  dickerIsMostExpensive: boolean;
}

export interface ComparisonResult {
  rows: ComparisonRow[];
  distributors: Distributor[];
}

export type DistributorCardFreshness = typeof DistributorCardFreshness[keyof typeof DistributorCardFreshness];


export const DistributorCardFreshness = {
  fresh: 'fresh',
  stale_warn: 'stale_warn',
  stale_critical: 'stale_critical',
  no_data: 'no_data',
} as const;

export interface DistributorCard {
  distributorId: number;
  name: string;
  isBaseline: boolean;
  stalenessThresholdDays: number;
  /** @nullable */
  lastUploadAt?: string | null;
  /** @nullable */
  lastUploadDate?: string | null;
  freshness: DistributorCardFreshness;
}

export interface DashboardSummary {
  distributorCards: DistributorCard[];
  totalProducts: number;
  dickerMostExpensiveCount: number;
  totalNetMovement: number;
}

export type ListUploadsParams = {
distributorId?: number;
};

export type GetComparisonParams = {
brand?: string;
distributorId?: number;
search?: string;
};