# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Distributor Pricing & Stock Comparison API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Email/password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Distributors ---
  /distributors:
    get:
      operationId: listDistributors
      tags: [distributors]
      summary: List all distributors
      responses:
        "200":
          description: List of distributors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distributor"

    post:
      operationId: createDistributor
      tags: [distributors]
      summary: Create a new distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistributorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /distributors/{id}:
    patch:
      operationId: updateDistributor
      tags: [distributors]
      summary: Update a distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistributorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteDistributor
      tags: [distributors]
      summary: Delete a distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Brand Aliases ---
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all canonical brands with their aliases
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"

    post:
      operationId: createBrand
      tags: [brands]
      summary: Create a new canonical brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [brands]
      summary: Update a brand and its aliases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"

    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete a brand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Import Profiles ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # --- Auth ---
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    # --- Distributors ---
    Distributor:
      type: object
      required: [id, name, isBaseline, stalenessThresholdDays, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        isBaseline:
          type: boolean
        stalenessThresholdDays:
          type: integer
        createdAt:
          type: string
        lastUploadAt:
          type: ["string", "null"]
        lastUploadStatus:
          type: ["string", "null"]
    DistributorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        isBaseline:
          type: boolean
        stalenessThresholdDays:
          type: integer
    DistributorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        isBaseline:
          type: boolean
        stalenessThresholdDays:
          type: integer

    # --- Brands ---
    Brand:
      type: object
      required: [id, canonicalName, aliases]
      properties:
        id:
          type: integer
        canonicalName:
          type: string
        aliases:
          type: array
          items:
            type: string
    BrandInput:
      type: object
      required: [canonicalName, aliases]
      properties:
        canonicalName:
          type: string
          minLength: 1
        aliases:
          type: array
          items:
            type: string
    BrandUpdate:
      type: object
      properties:
        canonicalName:
          type: string
          minLength: 1
        aliases:
          type: array
          items:
            type: string

    # --- Import Profiles ---
...