# lib/api-zod/src/generated/types/brandUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributor Pricing & Stock Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandUpdate {
  /** @minLength 1 */
  canonicalName?: string;
  aliases?: string[];
}