# lib/api-zod/src/generated/types/distributorUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributor Pricing & Stock Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface DistributorUpdate {
  /** @minLength 1 */
  name?: string;
  isBaseline?: boolean;
  stalenessThresholdDays?: number;
}